MAR 17

+++ semestru 4/mpp/ProiectMpp/ProiectMpp/Domain/Team.cs
@@ -1,6 +1,6 @@
 namespace ProiectMpp.Domain;
 
-public class Team : Entity<long>
+public class Team : Entity<int>
 {
     public string Name { get; set; }
